PHC/hw2/requirements.txt
onnxruntime
pytelegrambotapi
pillow
telegram
numpy
opencv-python
